requirements.txt
pinecone_client==2.2.2
sentence_transformers==2.2.2
streamlit==1.16.0
torch==2.0.0
tqdm==4.65.0
transformers==4.27.4
